refactor: simplify the account dropdown menu toggling logic (FE-88)
